+++ tests/integration/query_range_test.go
@@ -61,9 +61,9 @@ func TestQueryRange(t *testing.T) {
 		)
 		require.NoError(t, err)
 		count := 0
-		for rs, err := range r.Range(ctx) {
+		for rs, err := range r.ResultSets(ctx) {
 			require.NoError(t, err)
-			for row, err := range rs.Range(ctx) {
+			for row, err := range rs.Rows(ctx) {
 				require.NoError(t, err)
 
 				var (
@@ -109,11 +109,11 @@ func TestQueryRange(t *testing.T) {
 			if err != nil {
 				return err
 			}
-			for rs, err := range r.Range(ctx) {
+			for rs, err := range r.ResultSets(ctx) {
 				if err != nil {
 					return err
 				}
-				for row, err := range rs.Range(ctx) {
+				for row, err := range rs.Rows(ctx) {
 					if err != nil {
 						return err
 					}
@@ -133,7 +133,7 @@ func TestQueryRange(t *testing.T) {
 					}
 				}
 			}
-			return r.Err()
+			return nil
 		}, query.WithIdempotent())
 		require.NoError(t, err)
 	})
@@ -162,11 +162,11 @@ func TestQueryRange(t *testing.T) {
 			if err != nil {
 				return err
 			}
-			for rs, err := range r.Range(ctx) {
+			for rs, err := range r.ResultSets(ctx) {
 				if err != nil {
 					return err
 				}
-				for row, err := range rs.Range(ctx) {
+				for row, err := range rs.Rows(ctx) {
 					if err != nil {
 						return err
 					}
@@ -186,7 +186,7 @@ func TestQueryRange(t *testing.T) {
 					}
 				}
 			}
-			return r.Err()
+			return nil
 		}, query.WithIdempotent())
 		require.NoError(t, err)
 	})